:construction: Add hex color to sphere and image drawing

+++ src/geometry/circle.zig
@@ -1,7 +1,7 @@
 const Circle = @import("../datatypes/circle.zig");
 const std = @import("std");
 
-pub fn drawCircle(image: *[][]u8, circle: Circle.Circle) void {
+pub fn drawCircle(image: *[][]u32, circle: Circle.Circle) void {
     for (image.*, 0..) |row, index_column| {
         for (row, 0..) |_, index_row| {
             if (IsInsideCircle(circle, @as(i64, @intCast(index_column)), @as(i64, @intCast(index_row)))) {
